Update text tests
