+++ property_tests/tests/text/test/generate_tests.ts
@@ -1,0 +1,35 @@
+import { deepEqual } from 'fast-equals';
+
+import { getActor } from 'azle/property_tests';
+import { CandidMeta } from 'azle/property_tests/arbitraries/candid/candid_arb';
+import { Named } from 'azle/property_tests/arbitraries/query_method_arb';
+import { Test } from 'azle/test';
+
+export function generateTests(
+    functionName: string,
+    namedParamTexts: Named<CandidMeta<string>>[],
+    returnTexts: CandidMeta<string>
+): Test[] {
+    const expectedResult = namedParamTexts.reduce(
+        (acc, param) => acc + param.el.agentResponseValue,
+        returnTexts.agentResponseValue
+    );
+    const paramValues = namedParamTexts.map(
+        (param) => param.el.agentArgumentValue
+    );
+
+    return [
+        {
+            name: `text ${functionName}`,
+            test: async () => {
+                const actor = getActor('./tests/text/test');
+
+                const result = await actor[functionName](...paramValues);
+
+                return {
+                    Ok: deepEqual(result, expectedResult)
+                };
+            }
+        }
+    ];
+}
